save parcel data
